feat: on string_encoder, made a way to encode bytes in a more efficient way than base32768
feat: updated the order of parameter in bytes. it fucked everythng and it might contain some bugs now. still fixing

feat: added support for rollup with terser

organization: organized the code to be more modular, and organized

feat: added a constant for the max datagram size, which now is 2kb